intents: the watchdog's wait is milliseconds and milliseconds are a long
The deadline handed to the handler was already computed as a long; the watchdog's own wait
was not. Past 2147483 seconds -- a budget the annotation processor accepts and
setDefaultTimeout accepts -- the int multiplication overflowed to a negative, so
awaitCompletion reported a timeout at once. The platform was told the action had been
stopped while the handler was still running its side effects, and the context that handler
held said it had weeks left.

Pinned by a declaration with such a budget and a handler that takes 200ms: the result has
to be the handler's. The check is on the arithmetic rather than the declared type --
storing the product in a long while multiplying two ints still fails it, which is the
version of this bug that survives a careless fix.

+++ CodenameOne/src/com/codename1/intents/Intents.java
@@ -2081,7 +2081,12 @@ public final class Intents {
         final IntentContext ctx = new IntentContext(inv.source, inv.headless,
                 System.currentTimeMillis() + timeout * 1000L);
         final CompletionGuard guard = new CompletionGuard(inv.completion);
-        final int timeoutMillis = timeout * 1000;
+        // A long, like the deadline two lines above. As an int this overflowed for any budget
+        // past 2147483 seconds -- which the build accepts and setDefaultTimeout accepts -- and
+        // a negative wait made awaitCompletion report a timeout immediately: the platform was
+        // told the action was stopped while the handler was still running its side effects, and
+        // the context it held said it had weeks left.
+        final long timeoutMillis = timeout * 1000L;
 
         Runnable body = new Runnable() {
             @Override
